Update `DAG_to_restricted` using the new representation of B-sets
TODO: complete and debug this function.
This is not finished yet.

It can be optimized further, but note that completing all the interfering v-structures may not be as simple: changes at a given level may have consequences below.

It is better to do it node by node (and recompute the B-sets each time).